fix more numpy dtypes (#672)
* replace np.int by np.int64

* replace np.float by np.float64

* replace np.complex by np.complex128

+++ src/toast/fod/psd_math.py
@@ -135,7 +135,7 @@ def crosscov_psd(
     # for the unflagged samples
     naverage = lagmax
 
-    nreal = np.int(np.ceil((time_stop - time_start) / stationary_period))
+    nreal = np.int64(np.ceil((time_stop - time_start) / stationary_period))
 
     # Communicate lagmax samples from the beginning of the array
     # backwards in the MPI communicator
@@ -222,7 +222,7 @@ def crosscov_psd(
     # Collect the estimated covariance functions
 
     my_covs = {}
-    nreal_task = np.int(np.ceil(nreal / ntask))
+    nreal_task = np.int64(np.ceil(nreal / ntask))
 
     if comm is None:
         for ireal in range(nreal):
